refactor: FameGameFactory를 테스트 패키지로 이동
